// src/main/resources/www2/src/api/api.ts
export interface DeviceCredentialSet {
    /**
     * 
     * @type {number}
     * @memberof DeviceCredentialSet
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof DeviceCredentialSet
     */
    'name'?: string;
    /**
     * 
     * @type {Domain}
     * @memberof DeviceCredentialSet
     */
    'mgmtDomain'?: Domain;
    /**
     * 
     * @type {boolean}
     * @memberof DeviceCredentialSet
     */
    'deviceSpecific'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof DeviceCredentialSet
     */
    'type': string;
}
/**
 * 
 * @export
 * @interface DeviceCredentialSetRestApiView
 */
export interface DeviceCredentialSetRestApiView {
    /**
     * 
     * @type {number}
     * @memberof DeviceCredentialSetRestApiView
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof DeviceCredentialSetRestApiView
     */
    'name'?: string;
    /**
     * 
     * @type {DomainRestApiView}
     * @memberof DeviceCredentialSetRestApiView
     */
    'mgmtDomain'?: DomainRestApiView;
    /**
     * 
     * @type {boolean}
     * @memberof DeviceCredentialSetRestApiView
     */
    'deviceSpecific'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof DeviceCredentialSetRestApiView
     */
    'type': string;
}
// ...
export interface DeviceSnmpv1Community extends DeviceCredentialSet {
    /**
     * 
     * @type {string}
     * @memberof DeviceSnmpv1Community
     */
    'community'?: string;
}
/**
 * 
 * @export
 * @interface DeviceSnmpv1CommunityAllOf
 */
export interface DeviceSnmpv1CommunityAllOf {
    /**
     * 
     * @type {string}
     * @memberof DeviceSnmpv1CommunityAllOf
     */
    'community'?: string;
}
/**
 * 
 * @export
 * @interface DeviceSnmpv2cCommunity
 */
export interface DeviceSnmpv2cCommunity extends DeviceCredentialSet {
    /**
     * 
     * @type {string}
     * @memberof DeviceSnmpv2cCommunity
     */
    'community'?: string;
}
/**
 * 
 * @export
 * @interface DeviceSnmpv3Community
 */
export interface DeviceSnmpv3Community extends DeviceCredentialSet {
    /**
     * 
     * @type {string}
     * @memberof DeviceSnmpv3Community
     */
    'community'?: string;
    /**
     * 
     * @type {string}
     * @memberof DeviceSnmpv3Community
     */
    'username'?: string;
    /**
     * 
     * @type {string}
     * @memberof DeviceSnmpv3Community
     */
    'authType'?: string;
    /**
     * 
     * @type {string}
     * @memberof DeviceSnmpv3Community
     */
    'authKey'?: string;
    /**
     * 
     * @type {string}
     * @memberof DeviceSnmpv3Community
     */
    'privType'?: string;
    /**
     * 
     * @type {string}
     * @memberof DeviceSnmpv3Community
     */
    'privKey'?: string;
}
// ...
export interface DeviceSshAccount extends DeviceCredentialSet {
    /**
     * 
     * @type {string}
     * @memberof DeviceSshAccount
     */
    'username'?: string;
    /**
     * 
     * @type {string}
     * @memberof DeviceSshAccount
     */
    'password'?: string;
    /**
     * 
     * @type {string}
     * @memberof DeviceSshAccount
     */
    'superPassword'?: string;
}
// ...
export interface DeviceSshKeyAccount extends DeviceCredentialSet {
    /**
     * 
     * @type {string}
     * @memberof DeviceSshKeyAccount
     */
    'username'?: string;
    /**
     * 
     * @type {string}
     * @memberof DeviceSshKeyAccount
     */
    'password'?: string;
    /**
     * 
     * @type {string}
     * @memberof DeviceSshKeyAccount
     */
    'superPassword'?: string;
    /**
     * 
     * @type {string}
     * @memberof DeviceSshKeyAccount
     */
    'publicKey'?: string;
    /**
     * 
     * @type {string}
     * @memberof DeviceSshKeyAccount
     */
    'privateKey'?: string;
}
// ...
export interface DeviceTelnetAccount extends DeviceCredentialSet {
    /**
     * 
     * @type {string}
     * @memberof DeviceTelnetAccount
     */
    'username'?: string;
    /**
     * 
     * @type {string}
     * @memberof DeviceTelnetAccount
     */
    'password'?: string;
    /**
     * 
     * @type {string}
     * @memberof DeviceTelnetAccount
     */
    'superPassword'?: string;
}
// ...
export interface Domain {
    /**
     * 
     * @type {number}
     * @memberof Domain
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof Domain
     */
    'changeDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof Domain
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof Domain
     */
    'description'?: string;
    /**
     * 
     * @type {Network4Address}
     * @memberof Domain
     */
    'server4Address'?: Network4Address;
    /**
     * 
     * @type {Network6Address}
     * @memberof Domain
     */
    'server6Address'?: Network6Address;
}
/**
 * 
 * @export
 * @interface DomainRestApiView
 */
export interface DomainRestApiView {
    /**
     * 
     * @type {number}
     * @memberof DomainRestApiView
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof DomainRestApiView
     */
    'changeDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof DomainRestApiView
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof DomainRestApiView
     */
    'description'?: string;
    /**
     * 
     * @type {Network4AddressRestApiView}
     * @memberof DomainRestApiView
     */
    'server4Address'?: Network4AddressRestApiView;
    /**
     * 
     * @type {Network6AddressRestApiView}
     * @memberof DomainRestApiView
     */
    'server6Address'?: Network6AddressRestApiView;
}
// ...
export interface Network4AddressRestApiView {
    /**
     * 
     * @type {number}
     * @memberof Network4AddressRestApiView
     */
    'prefixLength'?: number;
    /**
     * 
     * @type {string}
     * @memberof Network4AddressRestApiView
     */
    'addressUsage'?: Network4AddressRestApiViewAddressUsageEnum;
    /**
     * 
     * @type {string}
     * @memberof Network4AddressRestApiView
     */
    'ip'?: string;
}
// ...
export interface Network6AddressRestApiView {
    /**
     * 
     * @type {number}
     * @memberof Network6AddressRestApiView
     */
    'prefixLength'?: number;
    /**
     * 
     * @type {string}
     * @memberof Network6AddressRestApiView
     */
    'addressUsage'?: Network6AddressRestApiViewAddressUsageEnum;
    /**
     * 
     * @type {string}
     * @memberof Network6AddressRestApiView
     */
    'ip'?: string;
}